Update HealComm-1.0.lua
Remove print statement. Probably debug print that stayed in.

Printed "HealComm predicts ..." in chat every time a heal is cast. I think this probably came in as a debug print and got forgotten. I didn't find a way to turn it off ingame, so I went to the code. As far as I can tell there is no way to turn it off other than deleting this line and as far as I know the print does not have any other effects.